Move helper method
